:heavy_plus_sign: Add getAddresses
