+++ FlipConnectSDK/FCConsts.swift
@@ -11,4 +11,5 @@ import Foundation
 struct FCConsts {
     static let connectWebUrl = "http://flipconnect-signin-staging.herokuapp.com/"
     static let connectApiUrl = "http://dlp-qrservices.cloudapp.net:20112/api/"
+    static let connectUserManagementUrl = "http://dlp-qrservices.cloudapp.net:20115/"
 }
